+++ lib/app/app_router/scaffold_with_nested_navigation.dart
@@ -38,7 +38,9 @@ class ScaffoldWithNestedNavigation extends StatelessWidget {
             ),
           ),
           BottomNavigationBarItem(
-              label: 'Settings', icon: Icon(Icons.settings)),
+            label: 'Settings',
+            icon: Icon(Icons.settings),
+          ),
         ],
         onTap: _goBranch,
       ),
